+++ src/main/java/org/mbari/mxm/provider_client/rest/MxmInfo.java
@@ -6,6 +6,5 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class MxmInfo {
 
-  // TODO or just a base endpoint on which to append particular routes depending on operation
-  public String missionStatusEndpoint;
+  public String mxmRestEndpoint;
 }