+++ src/main/java/org/mbari/mxm/provider_client/rest/PostMissionRequest.java
@@ -4,8 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 import lombok.AllArgsConstructor;
 
-public class PostMissionPayload {
+public class PostMissionRequest {
 
+  public String providerId;
   public String missionTplId;
   public String missionId;
   public String assetId;